JAVADOC DEPLOY: removed logic to move files
No longer checking out the other branch and attempting to move their files out of the folder due to build errors. If we want to go this route a better understanding of github workflows is required and should be obtained on a local system so that admins no longer get notified every time a commit is made to main.