fixed auton not ending
The issue was that all of robot code was hung on a single thread. This thread that it was hung on was the one running auton command's execute. When the timer reached a point where the last marker was behind it the bin search would never end. Better logic for bin search should be added but the short term solution that this commit implemented was setting the last marker to be 100 * the final pose in the trajectory. Breaks optimization methods.