added logic for continously checking if the driver station is connected in order to wait for it to connect to update the way that the trajectorys are transformed
the reason that we need this change is because when we connect to the robot with a computer it waits for a device to connect to it before it starts robotInit. This allows for driverstation to connect and therefore an alliance to be configured before we call certain methods for device setup. WE SHOULD ALSO ADD LOGIC TO CHANGE TRAJECTORIES TO BE RE-GENERATED WHEN THE BOT KNOWS WHAT ALLIANCE IT IS ON